Added timeout support to the requests we send
Unfortunately, the requests library swallows the `asyncio.TimeoutError` exception, therefore we can't kill the stop the thread early. Added a timeout to the request should fix this.
TODO: A new problem arises when we need to send multiple requests, each one will have the same timeout parameter, which means the total timeout isn't the requested timeout (the thread might not stop in time).